requirements.txt
pandas==3.0.1
numpy==2.4.3
scikit-learn==1.8.0
matplotlib==3.10.8
fastapi==0.128.0
polaris==0.1
sentence-transformers==5.3.0
youtube-transcript-api==1.2.4
requests==2.31.0
python-dotenv==1.1.0
sqlalchemy>=2.0
uvicorn[standard]
httpx>=0.27.0
